requirements.txt
flask
python-gedcom
pytest
ruff